monitoring/security: Hide delete comment action in the comments overview if user lacks the respective permission

+++ modules/monitoring/application/controllers/ListController.php
@@ -484,7 +484,10 @@ class Monitoring_ListController extends Controller
                 'comment_expiration'    => $this->translate('Expiration')
             )
         );
-        $this->view->delCommentForm = new DeleteCommentCommandForm();
+
+        if ($this->Auth()->hasPermission('monitoring/command/comment/delete')) {
+            $this->view->delCommentForm = new DeleteCommentCommandForm();
+        }
     }
 
     public function servicegroupsAction()
